+++ packages/react/src/context/index.tsx
@@ -1,24 +1,23 @@
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-// @ts-nocheck
 import React, { useMemo, createContext, useContext } from 'react';
 import Nile, { NileApi } from '@theniledev/js';
 import { CssVarsProvider } from '@mui/joy/styles';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 
-const queryClient = new QueryClient();
+import defaultTheme from './themeJoiner';
 import { NileContext, NileProviderProps } from './types';
+
+const defaultQueryClient = new QueryClient();
 
 const defaultContext: NileContext = {
   instance: Nile({ basePath: '', workspace: 'none', credentials: 'include' }),
-  workspace: '',
-  theme: {},
 };
 
 const context = createContext<NileContext>(defaultContext);
 const { Provider } = context;
 
 export const NileProvider = (props: NileProviderProps) => {
-  const { children } = props;
+  const { children, theme, queryClient } = props;
 
   const values = useMemo<NileContext>(() => {
     return {
@@ -27,23 +26,12 @@ export const NileProvider = (props: NileProviderProps) => {
         workspace: props.workspace,
         credentials: 'include',
       }),
-      theme: props.theme,
     };
-  }, [props.basePath, props.theme, props.workspace]);
-
-  if (props.theme) {
-    return (
-      <QueryClientProvider client={queryClient}>
-        <CssVarsProvider theme={props.theme}>
-          <Provider value={values}>{children}</Provider>
-        </CssVarsProvider>
-      </QueryClientProvider>
-    );
-  }
+  }, [props.basePath, props.workspace]);
 
   return (
-    <QueryClientProvider client={queryClient}>
-      <CssVarsProvider>
+    <QueryClientProvider client={queryClient ?? defaultQueryClient}>
+      <CssVarsProvider theme={theme ?? defaultTheme}>
         <Provider value={values}>{children}</Provider>
       </CssVarsProvider>
     </QueryClientProvider>
@@ -56,8 +44,4 @@ const useNileContext = (): NileContext => {
 
 export const useNile = (): NileApi => {
   return useNileContext().instance;
-};
-
-export const useNileContextTheme = (): void | string => {
-  return useNileContext().theme;
 };