+++ packages/react/src/context/types.ts
@@ -1,6 +1,7 @@
 import { NileApi } from '@theniledev/js';
 import React from 'react';
 import { Theme } from '@mui/joy/styles';
+import { QueryClient } from '@tanstack/react-query';
 
 export interface NileContext {
   instance: NileApi;
@@ -11,5 +12,6 @@ export interface NileProviderProps {
   children: React.ReactNode;
   basePath: string;
   workspace?: string;
-  theme?: string;
+  theme?: Theme;
+  queryClient?: QueryClient;
 }