+++ packages/react/src/index.ts
@@ -1,4 +1,5 @@
 export { default as LoginForm } from './components/LoginForm';
 export { default as SignUpForm } from './components/SignUpForm';
-
+export { default as InstanceTable } from './components/InstanceTable';
 export { NileProvider, useNile } from './context';
+export { default as Queries, useQuery, useMutation } from './lib/queries';