+++ CHANGELOG.md
@@ -3,6 +3,89 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.1](https://github.com/TheNileDev/nile-js/compare/v0.16.0...v0.16.1) (2022-08-18)
+
+
+### Bug Fixes
+
+* **react:** support rendering unknown columns ([6a2ec0b](https://github.com/TheNileDev/nile-js/commit/6a2ec0bceae5d652dd7df3eda13fbcd6a93218eb))
+
+
+
+
+
+# [0.16.0](https://github.com/TheNileDev/nile-js/compare/v0.15.0...v0.16.0) (2022-08-18)
+
+
+### Bug Fixes
+
+* **react:** use correct names for user form ([d70a9ef](https://github.com/TheNileDev/nile-js/commit/d70a9ef2edd36bbb6c0bb46e0a5cb277f063d610))
+
+
+### Features
+
+* **react:** export useQuery and useMutation ([0f229e4](https://github.com/TheNileDev/nile-js/commit/0f229e4901e25d63c013ad42531bc98b71cad881))
+
+
+
+
+
+# [0.15.0](https://github.com/TheNileDev/nile-js/compare/v0.14.2...v0.15.0) (2022-08-17)
+
+
+### Bug Fixes
+
+* **react:** add more padding height to instance table ([d73a16d](https://github.com/TheNileDev/nile-js/commit/d73a16daf53ba44bb25f7c9ac7df43ad65f11ffd))
+
+
+### Features
+
+* **react:** allow a custom react-query queryclient ([3e3da5a](https://github.com/TheNileDev/nile-js/commit/3e3da5ab50ef3cd6917c7aa26ccc50d39bb96a05))
+
+
+
+
+
+## [0.14.2](https://github.com/TheNileDev/nile-js/compare/v0.14.1...v0.14.2) (2022-08-17)
+
+
+### Bug Fixes
+
+* expect a full config object when provided ([0565f30](https://github.com/TheNileDev/nile-js/commit/0565f30ad19b32dd91462aa52e6a73e7038bd64f))
+
+
+
+
+
+## [0.14.1](https://github.com/TheNileDev/nile-js/compare/v0.14.0...v0.14.1) (2022-08-17)
+
+**Note:** Version bump only for package @theniledev/packages
+
+
+
+
+
+# [0.14.0](https://github.com/TheNileDev/nile-js/compare/v0.13.1...v0.14.0) (2022-08-16)
+
+
+### Bug Fixes
+
+* clean up deployment code ([e585045](https://github.com/TheNileDev/nile-js/commit/e585045496da2062869bb58e8ccf12b49801a62c))
+* correct pulumi project name and clarify reconciliation plan ([fd00feb](https://github.com/TheNileDev/nile-js/commit/fd00feb63b65ba872077bea58054a7fa071cc4ae))
+* **events-example:** linting on events, documentation ([fa464e5](https://github.com/TheNileDev/nile-js/commit/fa464e5a77aedd987aa996b9d47c571908f62ce3))
+* **nile-js:** add typedoc to events class ([e4f66e8](https://github.com/TheNileDev/nile-js/commit/e4f66e8f495d4936432f6a0fb00ed82125ee4bdb))
+* remove commented code from reconciler ([952c86d](https://github.com/TheNileDev/nile-js/commit/952c86dc31743260853724174156c1a5f9830878))
+
+
+### Features
+
+* add instance table ([ed0a057](https://github.com/TheNileDev/nile-js/commit/ed0a057856a88c11797f99c3a6e8955905f478dc))
+* **react:** have opinions about libs, upgrade to use them ([498af4e](https://github.com/TheNileDev/nile-js/commit/498af4e80b0dc880488c14aa89062acd5cac652d))
+
+
+
+
+
 ## [0.13.1](https://github.com/TheNileDev/nile-js/compare/v0.13.0...v0.13.1) (2022-08-02)