+++ lib/nile/CHANGELOG.md
@@ -3,6 +3,58 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.1](https://github.com/TheNileDev/nile-js/compare/v0.16.0...v0.16.1) (2022-08-18)
+
+**Note:** Version bump only for package @theniledev/js
+
+
+
+
+
+# [0.16.0](https://github.com/TheNileDev/nile-js/compare/v0.15.0...v0.16.0) (2022-08-18)
+
+**Note:** Version bump only for package @theniledev/js
+
+
+
+
+
+# [0.15.0](https://github.com/TheNileDev/nile-js/compare/v0.14.2...v0.15.0) (2022-08-17)
+
+**Note:** Version bump only for package @theniledev/js
+
+
+
+
+
+## [0.14.2](https://github.com/TheNileDev/nile-js/compare/v0.14.1...v0.14.2) (2022-08-17)
+
+**Note:** Version bump only for package @theniledev/js
+
+
+
+
+
+## [0.14.1](https://github.com/TheNileDev/nile-js/compare/v0.14.0...v0.14.1) (2022-08-17)
+
+**Note:** Version bump only for package @theniledev/js
+
+
+
+
+
+# [0.14.0](https://github.com/TheNileDev/nile-js/compare/v0.13.1...v0.14.0) (2022-08-16)
+
+
+### Bug Fixes
+
+* **events-example:** linting on events, documentation ([fa464e5](https://github.com/TheNileDev/nile-js/commit/fa464e5a77aedd987aa996b9d47c571908f62ce3))
+* **nile-js:** add typedoc to events class ([e4f66e8](https://github.com/TheNileDev/nile-js/commit/e4f66e8f495d4936432f6a0fb00ed82125ee4bdb))
+
+
+
+
+
 ## [0.13.1](https://github.com/TheNileDev/nile-js/compare/v0.13.0...v0.13.1) (2022-08-02)