+++ packages/examples/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.1](https://github.com/TheNileDev/nile-js/compare/v0.16.0...v0.16.1) (2022-08-18)
+
+**Note:** Version bump only for package @theniledev/examples
+
+
+
+
+
+# [0.16.0](https://github.com/TheNileDev/nile-js/compare/v0.15.0...v0.16.0) (2022-08-18)
+
+**Note:** Version bump only for package @theniledev/examples
+
+
+
+
+
+# [0.15.0](https://github.com/TheNileDev/nile-js/compare/v0.14.2...v0.15.0) (2022-08-17)
+
+**Note:** Version bump only for package @theniledev/examples
+
+
+
+
+
+## [0.14.2](https://github.com/TheNileDev/nile-js/compare/v0.14.1...v0.14.2) (2022-08-17)
+
+**Note:** Version bump only for package @theniledev/examples
+
+
+
+
+
+## [0.14.1](https://github.com/TheNileDev/nile-js/compare/v0.14.0...v0.14.1) (2022-08-17)
+
+**Note:** Version bump only for package @theniledev/examples
+
+
+
+
+
+# [0.14.0](https://github.com/TheNileDev/nile-js/compare/v0.13.1...v0.14.0) (2022-08-16)
+
+
+### Features
+
+* **react:** have opinions about libs, upgrade to use them ([498af4e](https://github.com/TheNileDev/nile-js/commit/498af4e80b0dc880488c14aa89062acd5cac652d))
+
+
+
+
+
 ## [0.13.1](https://github.com/TheNileDev/nile-js/compare/v0.13.0...v0.13.1) (2022-08-02)
 
 **Note:** Version bump only for package @theniledev/examples